+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -249,6 +249,11 @@ New check aliases
   <clang-tidy/checks/bugprone/throwing-static-initialization>`
   keeping initial check as an alias to the new one.
 
+- Renamed :doc:`cert-err60-cpp <clang-tidy/checks/cert/err60-cpp>` to
+  :doc:`bugprone-exception-type-not-nothrow-copy-constructible
+  <clang-tidy/checks/bugprone/exception-type-not-nothrow-copy-constructible>`
+  keeping initial check as an alias to the new one.
+
 Changes in existing checks
 ^^^^^^^^^^^^^^^^^^^^^^^^^^